Connect to localhost database
It was a struggle to connect to localhost db...
I've installed latest docker -v "18.9.0 test", newest docker-compose 1.23.1 and set up container network_mode to host.

It works now but it is not recommended way of manipulating with docker container network and that is clearly stated in their docs.

There is a solution in windows and mac systems which works by just adding host=host.docker.internal in databases connector, but from ubuntu version 18 it doesn't work on linux system.